cr-256: add support for provision agreement and move test to 3.0 folder
