+++ src/pysdmx/io/xml/sdmx21/__tokens.py
@@ -28,6 +28,7 @@ STR_REF = "structureRef"
 STRUCTURE = "Structure"
 STR_USAGE = "StructureUsage"
 PROV_AGREMENT = "ProvisionAgrement"
+PROV_AGREEMENT = "ProvisionAgreement"
 STR_ID = "structureID"
 STR_TYPE = "structure_type"
 DIM_OBS = "dimensionAtObservation"
