+++ src/pysdmx/io/xml/sdmx21/reader/header.py
@@ -16,6 +16,7 @@ from pysdmx.io.xml.sdmx21.__tokens import (
     ID,
     NAME,
     PREPARED,
+    PROV_AGREEMENT,
     PROV_AGREMENT,
     RECEIVER,
     REF,
@@ -31,7 +32,7 @@ from pysdmx.io.xml.sdmx21.__tokens import (
 from pysdmx.io.xml.sdmx21.reader.__parse_xml import parse_xml
 from pysdmx.model import Organisation, Reference
 from pysdmx.model.message import Header
-from pysdmx.util import parse_maintainable_urn, parse_urn
+from pysdmx.util import parse_maintainable_urn
 
 
 def __parse_sender_receiver(
@@ -96,6 +97,9 @@ def __parse_structure(
     elif STR_USAGE in structure:
         structure_info = structure[STR_USAGE]
         sdmx_type = DFW
+    elif PROV_AGREEMENT in structure:
+        structure_info = structure[PROV_AGREEMENT]
+        sdmx_type = PROV_AGREEMENT
     else:
         structure_info = structure[PROV_AGREMENT]
         sdmx_type = PROV_AGREMENT
@@ -114,7 +118,7 @@ def __parse_structure(
     elif URN in structure_info:
         ref_obj = parse_maintainable_urn(structure_info[URN])
     else:
-        ref_obj = parse_urn(structure_info)
+        ref_obj = parse_maintainable_urn(structure_info)
     return {str(ref_obj): dim_at_obs}
